requirements.txt
python-telegram-bot==20.8
mega.py==1.0.8
python-dotenv==1.0.0